+++ engine/src/main/kotlin/engine/expressions/VariableExpression.kt
@@ -16,19 +16,7 @@ class Variable(
 ) {
     override fun signOf() = Sign.UNKNOWN
 
-    override fun toJson(): List<Any> {
-        return if (decorators.isEmpty() && name == null) {
-            listOf(operator.toString())
-        } else {
-            listOf(
-                listOf(operator.toString()) +
-                    (if (name == null) listOf() else listOf(name)) +
-                    decorators.map { it.toString() },
-            )
-        }
-    }
-
-    override fun fillJson2(s: MutableMap<String, Any>) {
+    override fun fillJson(s: MutableMap<String, Any>) {
         s["type"] = "Variable"
         s["value"] = variableName
         if (subscript != null) {
@@ -48,7 +36,7 @@ class Name(
     val text: String,
     meta: NodeMeta = BasicMeta(),
 ) : Expression(NameOperator(text), listOf(), meta) {
-    override fun fillJson2(s: MutableMap<String, Any>) {
+    override fun fillJson(s: MutableMap<String, Any>) {
         s["type"] = "Name"
         s["value"] = text
     }